Show PAT nag message, stop signout message
We're removing support for PAT tokens in user settings so if we find one, send telemetry about it, ignore it, and display a warning message to the user.  Only show the warning (nag) message once (on initialization).  If re-initialization occurs, don't show the nag message again.  If VS Code is restarted and the token is still present, then we'll show the nag message again.

On signout, reinitialization happens.  When it does, we always show the message that there's no credentials stored for the account (duh; we just removed them).  So prevent that message from being shown when re-init happens via signout.